lint: expand ruff rules (SIM/C4/PT/RUF), bulk format, strict CI
pyproject.toml:
  - Added SIM, C4, PT, RUF rule sets to ruff lint select
  - Added per-file-ignores for E402 (post-sys.path imports)
  - Ignored noisy rules: SIM102/108/115/117, RUF005/012/015, etc.

.github/workflows/ci.yml:
  - Ruff format check now strict (removed continue-on-error)

Auto-fixed: RUF046, RUF100, C401, C403, PT006, SIM210, SIM118
Manual fixes: noqa for SCons globals, random.randint, import-tests

All 6474 Python tests pass, ruff check + format clean.

+++ cyberpunk_designer/selection_ops/layout.py
@@ -241,7 +241,7 @@ def grid_arrange(app) -> None:
     except Exception:
         pass
     n = len(valid)
-    cols = max(1, int(math.ceil(math.sqrt(n))))
+    cols = max(1, math.ceil(math.sqrt(n)))
     # Use first widget position as origin; uniform cell = max sizes + GRID gap
     max_w = max(int(sc.widgets[i].width or GRID) for i in valid)
     max_h = max(int(sc.widgets[i].height or GRID) for i in valid)
